Changes used algorithm from depth-first to breadth-first
It's a failed attempt to optimize the algorithm, but it's worth
to be recorded. Breadth-first search is much less memory-efficient
then depth-first search, so all the attempts to run the tests
resulted in OutOfMemoryException. I could've changed maximum heap
size, but that wouldn't make it much faster, because of all
the data stored as a potential result. It lands in tenured space
pretty fast and cannot be collected before the end of computation,
hence the error.