Provide a cut-off for search tree
While doing a search on a tree the algorithm enters
the same paths in different permutations. Previous solution
was to sort figures lists based on toString output and filter them
using Set. Now it records successful paths and filters them
based on investigated figure type. If it already found solution with
quens placed on (1,1) and (2, 3), it will not further investigate (1, 1)
sub tree under (2, 3) node.

I've also tested changing Set to List as it's no longer needed, but it
resulted in a visible performance loss (about 20%)

Execution time for this solution is 27 seconds.