Adds filtering of figures variations
In this kind of problems solutions set can be created out of
certain number of fundamental solutions by combination of
rotations and reflections. This commit changes algorithm output
to quasi-fundamental solutions, which than can be unfolded into
list of all possible solutions.

There are two problems with this solutions:
- The resulting solutions are only quasi-fundamental. Algorithm
filters variations of a figure placed first on the board and filters
variations of next figures, if they do not change places of the
rest of the figures. It doesn't filter solutions, that are variations
of some other complete solutions.
- Variations generations turned out to be pretty expensive
task. While fundamental solutions can be generated in under 10 seconds,
the rest of the solutions are then generated in about a minute.